Dictionary of Active Learning Distances

A simple mlr3misc::Dictionary storing objects of class ALDistance. Each distance has an associated help page, see mlr_al_distances_[id].

For a more convenient way to retrieve and construct a distance, see clx_ald() and clx_alds().

Format

R6::R6Class object inheriting from mlr3misc::Dictionary.

Methods

See mlr3misc::Dictionary.

See also

Examples

as.data.table(mlr_al_distances)
#> Key: <key>
#>            key                    label                                  man
#>         <char>                   <char>                               <char>
#> 1:      affine       Affine AL Distance      celecx::mlr_al_distances_affine
#> 2:       gower        Gower AL Distance       celecx::mlr_al_distances_gower
#> 3: standardize Standardized AL Distance celecx::mlr_al_distances_standardize
clx_ald("standardize")
#> <ALDistanceStandardize:standardize>