Dictionary of Space Samplers

A simple mlr3misc::Dictionary storing objects of class SpaceSampler. Each sampler has an associated help page, see mlr_space_samplers_[id].

For a more convenient way to retrieve and construct a sampler, see clx_sps() and clx_spss().

Format

R6::R6Class object inheriting from mlr3misc::Dictionary.

Methods

See mlr3misc::Dictionary.

See also

Examples

as.data.table(mlr_space_samplers)
#> Key: <key>
#>                  key                            label
#>               <char>                           <char>
#> 1:             chain            Chained Space Sampler
#> 2:       conditional        Conditional Space Sampler
#> 3:               gsx                GSx Space Sampler
#> 4:            kmeans            K-Means Space Sampler
#> 5:          kmedoids          K-Medoids Space Sampler
#> 6:               lhs                LHS Space Sampler
#> 7: relational_kmeans Relational K-Means Space Sampler
#> 8:             sobol              Sobol Space Sampler
#> 9:           uniform            Uniform Space Sampler
#>                                             man deterministic
#>                                          <char>        <lgcl>
#> 1:             celecx::mlr_space_samplers_chain          TRUE
#> 2:       celecx::mlr_space_samplers_conditional          TRUE
#> 3:               celecx::mlr_space_samplers_gsx         FALSE
#> 4:            celecx::mlr_space_samplers_kmeans         FALSE
#> 5:          celecx::mlr_space_samplers_kmedoids          TRUE
#> 6:               celecx::mlr_space_samplers_lhs         FALSE
#> 7: celecx::mlr_space_samplers_relational_kmeans         FALSE
#> 8:             celecx::mlr_space_samplers_sobol         FALSE
#> 9:           celecx::mlr_space_samplers_uniform          TRUE
clx_sps("uniform")
#> <SpaceSamplerUniform:uniform>